fix: component tests
