+++ src/test/components/tests/test_execution_tags.py
@@ -1,34 +1,43 @@
 import unittest
-from test.test_utils.span_exporter import wait_for_exporter
 from test.test_utils.spans_parser import SpansContainer
 
 import requests
 
+from .app_runner import FastApiSample
+
 
 class TestExecutionTags(unittest.TestCase):
     def test_execution_tag(self):
-        response = requests.get("http://localhost:8020/invoke-request")
-        response.raise_for_status()
+        with FastApiSample():
+            response = requests.get("http://localhost:8020/invoke-request")
+            response.raise_for_status()
 
-        body = response.json()
+            body = response.json()
 
-        assert body is not None
+            assert body is not None
 
-        wait_for_exporter()
+            spans_container = SpansContainer.get_spans_from_file(
+                wait_time_sec=10, expected_span_count=4
+            )
+            self.assertEqual(4, len(spans_container.spans))
 
-        spans_container = SpansContainer.get_spans_from_file()
-        self.assertEqual(4, len(spans_container.spans))
-
-        # assert root
-        root = spans_container.get_first_root()
-        self.assertIsNotNone(root)
-        root_attributes = root["attributes"]
-        self.assertEqual(
-            type(root_attributes["lumigo.execution_tags.response_len"]), int
-        )
-        self.assertEqual(type(root_attributes["lumigo.execution_tags.response"]), str)
-        self.assertEqual(root_attributes["lumigo.execution_tags.app"], True)
-        self.assertEqual(
-            root_attributes["lumigo.execution_tags.app.response"], "success"
-        )
-        self.assertEqual(root_attributes["lumigo.execution_tags.foo"], ["bar", "baz"])
+            # assert root
+            root = spans_container.get_first_root()
+            self.assertIsNotNone(root)
+            root_attributes = root["attributes"]
+            self.assertEqual(
+                type(root_attributes["lumigo.execution_tags.response_len"]), int
+            )
+            self.assertEqual(
+                type(root_attributes["lumigo.execution_tags.response"]), str
+            )
+            self.assertEqual(root_attributes["lumigo.execution_tags.app"], True)
+            self.assertEqual(
+                root_attributes["lumigo.execution_tags.app.response"], "success"
+            )
+            self.assertEqual(
+                root_attributes["lumigo.execution_tags.foo"], ["bar", "baz"]
+            )
+            self.assertEqual(
+                root_attributes["lumigo.execution_tags.foo"], ["bar", "baz"]
+            )
